feat: implement FeaturesContent component with hash navigation and structured layout
