+++ app/(public)/features/_components/FeatureSectionHeader.tsx
@@ -1,0 +1,8 @@
+export function FeatureSectionHeader() {
+  return (
+    <h2 className="text-2xl sm:text-3xl md:text-4xl lg:text-5xl text-center text-gray-900 mt-6 sm:mt-8 md:mt-10 mb-4 sm:mb-5 md:mb-6 font-playfair px-2 sm:px-0">
+      Everything You Need,{" "}
+      <span className="italic text-adult-green">In One Place</span>
+    </h2>
+  );
+}
